[fix] Fix wrong base type of enum class
1. enum class ReplayOpCodes from int8_t to int16_t.

2. enum class ChunkHeaderType from uint8_t to uint16_t.

+++ RagnarokReplayCpp/ChunkHeaderType.hpp
@@ -5,7 +5,7 @@
 
 namespace RagnarokReplay {
 
-enum class ChunkHeaderType : uint8_t {
+enum class ChunkHeaderType : uint16_t {
   HEADER_CA_LOGIN = 0x0064,
   HEADER_CH_ENTER = 0x0065,
   HEADER_CH_SELECT_CHAR = 0x0066,
